# backend/spec/openapi.yaml
openapi: 3.0.3
info:
  title: 昆山佰泰胜专属 ERP 系统 API
  description: |
    V1.3.7 完整契约 · 13 Epic / 220+ 端点

    **升级要点**（vs V1.3.1）：
    - 工序分配职责严格分离生管 vs 采购（AD-1）
    - 月度对账不含"线下"动作（AD-2）
    - 单一 163 邮箱通知渠道（AD-3）
    - 委外 7 状态机 + 返修闭环 + 历史交期预估（V1.3.4）
    - 料号 5 段成本聚合（V1.3.4）
    - 签字扫描件 AES-256-GCM 加密（V1.3.6）

    **配套文档**：
    - `docs/prd.md` V1.3.7
    - `docs/architect-handoff.md` V1.1
    - `docs/ux-handoff.md` V1.1
    - `backend/spec/openapi.yaml`（本文件）
  version: 1.3.7
  contact:
    name: 河南晓评信息科技有限公司
    email: panqiang@yourcompany.com
  license:
    name: Proprietary
servers:
  - url: https://erp.yourcompany.local
    description: 生产环境（甲方内网）
  - url: http://localhost:8080
    description: 本地开发（通过 erp-gateway 8080）
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT (HS256) · access 2h + refresh 7d
        Header: `Authorization: Bearer <token>`

  # ---------- Schemas ----------
  schemas:
...
    Result:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: {}
        traceId: { type: string, example: abc123 }
...
    Order:
      type: object
      description: 'V1.3.7 Story 1.6 订单主表（crm_order）'
      properties:
        id: { type: integer }
        orderNo: { type: string, example: 'XS20260612-0001', description: 'XS+YYYYMMDD+NNNN（继承 1.5 DocNoGenerator）' }
        quoteId: { type: integer, description: '来源报价ID(1.5 转单时非空)' }
        customerId: { type: integer }
        customerName: { type: string }
        ownerUserId: { type: integer }
        deptId: { type: integer }
        currency: { type: string, default: 'CNY' }
        totalAmount: { type: number, format: double, description: '由 items 自动计算（只读）' }
        deliveryDate: { type: string, format: date }
        isFa: { type: boolean }
        isNew: { type: boolean }
        isUrgent: { type: boolean, description: 'V1.3.7 加急' }
        status:
          type: string
          enum: [DRAFT, CONFIRMED, PRODUCING, PARTIAL_SHIPPED, SHIPPED, SETTLED, CLOSED, CANCELLED]
          description: 'V1.3.7 §附录-b 7 状态机'
        currentNode: { type: integer, description: '当前审批节点 1/2/3' }
        comment: { type: string }
        productionOrderNo: { type: string, example: 'GD20260612-0001', description: '转生产时生成（Epic 5）' }
        outsourceOrderNo: { type: string, example: 'WW20260612-0001', description: '转委外时生成（Epic 6）' }
        creditLimitCheck: { type: integer, description: '0=未通过, 1=通过, -1=无限制' }
        isDeleted: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OrderItem:
      type: object
      description: '订单明细（crm_order_item，含 quantityAdjustment hook）'
      properties:
        id: { type: integer }
        orderId: { type: integer }
        drawingNo: { type: string, example: 'DWG-2026-001' }
        material: { type: string, example: '45#' }
        spec: { type: string, example: 'M10x20' }
        quantity: { type: integer }
        unitPrice: { type: number }
        amount: { type: number }
        quantityAdjustment: { type: integer, description: 'V1.3.7 数量调整（来自 1.5 hook）' }
        isFa: { type: boolean }
        isNew: { type: boolean }
        sort: { type: integer }
        producedQty: { type: integer, description: '已生产数量' }
        shippedQty: { type: integer, description: '已发货数量（累计）' }
...
    OrderUpdateRequest:
      type: object
      description: '订单更新请求（仅 DRAFT 状态可改）'
      properties:
        order: { $ref: '#/components/schemas/Order' }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
    OrderConfirmRequest:
      type: object
      description: '订单确认请求（DRAFT → CONFIRMED，> 20万 必填二次密码）'
      properties:
        secondPassword: { type: string, description: '> 20万 二次密码确认' }
        comment: { type: string }
...
    Reconcile:
      type: object
      description: 'V1.3.7 AD-2 红线：不含"线下"动作。状态机：DRAFT(1) → VENDOR_CONFIRMED(2) → BOTH_CONFIRMED(3) → FINANCE_CONFIRMED(4) → CLOSED'
      properties:
        id: { type: integer }
        reconcileNo: { type: string, example: 'RC202605-0001', description: 'RC{yyyyMM}{seq:4}（按月隔离）' }
        vendorId: { type: integer }
        vendorName: { type: string }
        periodYear: { type: integer, example: 2026 }
        periodMonth: { type: integer, example: 5, description: '1-12' }
        totalAmount: { type: number, example: 50000.00 }
        status: { type: string, enum: [DRAFT, VENDOR_CONFIRMED, BOTH_CONFIRMED, FINANCE_CONFIRMED, CLOSED] }
        currentStep: { type: integer, example: 1, description: '1-4 步' }
        isLocked: { type: boolean, description: '对账期锁定' }
        createdBy: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    # 对账明细（Story 1.21）
    ReconcileItem:
      type: object
      properties:
        id: { type: integer }
        reconcileId: { type: integer }
        outsourceOrderId: { type: integer }
        outsourceOrderNo: { type: string, example: 'WW20260612-0001' }
        itemName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        amount: { type: number }
        vendorAmount: { type: number, description: '厂商确认金额（AC-6.1.2）' }
        finalAmount: { type: number, description: '最终对账金额' }
        isConsistent: { type: boolean, description: '金额是否一致（40905 校验）' }
        sort: { type: integer }

    # 对账签字扫描件（Story 1.21 · AES-256-GCM）
    ReconcileSignature:
      type: object
      properties:
        id: { type: integer }
        reconcileId: { type: integer }
        signerUserId: { type: integer }
        signerName: { type: string }
        signatureImagePath: { type: string }
        encryptedData: { type: string, description: 'AES-256-GCM 加密 Base64' }
        iv: { type: string, description: '12 字节 IV 唯一（Base64）' }
        authTag: { type: string, description: '128-bit GCM tag（Base64）' }
        signedAt: { type: string, format: date-time }

    # 对账创建请求（Story 1.21）
...
    ReconcileItemRequest:
      type: object
      required: [outsourceOrderId, itemName, quantity, unitPrice]
      properties:
        outsourceOrderId: { type: integer }
        outsourceOrderNo: { type: string }
        itemName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }

    # 厂商对账确认请求（AC-6.1.2 · 40905）
...
paths:
...
  /quotes/{id}/submit:
    post:
      tags: [E2-Quote-Flow]
      summary: 提交审批（DRAFT → SUBMITTED，触发 stream:notify 4 通道推送）
      operationId: submitQuote
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功 }
        '40904': { description: QUOTE_STATE_INVALID（状态机越权） }
...
  /orders/{id}:
    get:
      tags: [E3-Order]
      summary: 查询订单详情（含状态机 + 历史 + 转生产/委外链路）
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功 }
        '40401': { description: ORDER_NOT_FOUND }
    put:
      tags: [E3-Order]
      summary: 修改订单（仅 DRAFT 状态可改）
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderUpdateRequest' }
      responses:
        '200': { description: 成功 }
        '40903': { description: ORDER_NOT_EDITABLE（仅 DRAFT 可改） }
  /orders/{id}/confirm:
    post:
      tags: [E3-Order-Flow]
      summary: 确认订单（DRAFT → CONFIRMED，触发 4 通道推送）
      operationId: confirmOrder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderConfirmRequest' }
      responses:
        '200': { description: 成功 }
        '40904': { description: ORDER_STATE_INVALID }
...
  /orders/{id}/close:
    post:
      tags: [E3-Order-Transfer]
      summary: 关闭（SETTLED → CLOSED 终态）
      operationId: closeOrder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功 }
        '40904': { description: ORDER_STATE_INVALID }
...
  /allocations:
    post:
      tags: [E5-Allocation]
      summary: '【生管】分配工序归属（自制/委外）'
      description: |
        **V1.3.7 红线**：此接口只能由生管调用，**不接受 vendorId 字段**。
        勾"委外"的工序会自动推送给采购的"待委外清单"。
      operationId: createAllocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workorderId, processSeq, decision]
              properties:
                workorderId: { type: integer }
                processSeq: { type: integer }
                decision: { type: string, enum: [INHOUSE, OUTSOURCE] }
                qty: { type: integer }
      responses:
        '200': { description: 成功 }
        '40304': { description: 工序分配越权（调用方非生管角色）' }
...
  /reconciles:
    post:
      tags: [E6-Reconcile]
      summary: 采购生成对账单 PDF
      description: '**V1.3.6 改版**：不含"采购带纸去厂商处"等线下动作'
      operationId: createReconcile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vendorId, period]
              properties:
                vendorId: { type: integer }
                period: { type: string, example: '2026-05' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Reconcile' }
...
  /reconciles/list:
    get:
      tags: [E6-Reconcile]
      summary: 对账列表（按 vendor/period/status 过滤）
      operationId: listReconciles
      parameters:
        - name: vendorId
          in: query
          schema: { type: integer }
        - name: periodYear
          in: query
          schema: { type: integer }
        - name: periodMonth
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string, enum: [DRAFT, VENDOR_CONFIRMED, BOTH_CONFIRMED, FINANCE_CONFIRMED, CLOSED] }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list: { type: array, items: { $ref: '#/components/schemas/Reconcile' } }
                          page: { type: integer }
                          size: { type: integer }
  /reconciles/{id}/detail:
    get:
      tags: [E6-Reconcile]
      summary: 对账单详情（主单 + 明细 + 签字）
      operationId: getReconcileDetail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          reconcile: { $ref: '#/components/schemas/Reconcile' }
                          items: { type: array, items: { $ref: '#/components/schemas/ReconcileItem' } }
                          signatures: { type: array, items: { $ref: '#/components/schemas/ReconcileSignature' } }
  /reconciles/{id}/items:
    post:
      tags: [E6-Reconcile]
      summary: 追加对账明细（AC-6.1.1）
      operationId: addReconcileItem
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReconcileItemRequest' }
      responses:
        '200': { description: 成功 }
        '40001': { description: 参数缺失 }
        '40404': { description: 对账单不存在 }
  /reconciles/{id}/upload-signature:
    post:
      tags: [E6-Reconcile]
      summary: 上传厂商签字扫描件（V1.3.7 AD-2 · 不含"线下" · AES-256-GCM 加密）
      description: |
        **V1.3.7 AD-2 红线**：通过电子方式上传，不允许线下传递
        **加密**：AES-256-GCM · IV 唯一 · 128-bit GCM tag
        **厂商签字必传**（P1 修补 3）
      operationId: uploadReconcileSignature
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, signerName]
              properties:
                file:
                  type: string
                  format: binary
                  description: 'PDF / JPG / PNG · ≤ 10MB'
                signerName:
                  type: string
                  description: '厂商签字人姓名'
      responses:
        '200':
          description: 成功（自动 AES-256-GCM 加密）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ReconcileSignature' }
  /reconciles/{id}/both-confirm:
    post:
      tags: [E6-Reconcile]
      summary: 双方对账确认 → step 3
      operationId: bothConfirmReconcile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功 }
        '40903': { description: 状态非法（需 VENDOR_CONFIRMED） }
  /reconciles/{id}/finance-confirm:
    post:
      tags: [E6-Reconcile]
      summary: 财务对账确认（AC-6.1.4 · step 4 → CLOSED）
      operationId: financeConfirmReconcile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功 → CLOSED + isLocked=true }
        '40903': { description: 状态非法（需 BOTH_CONFIRMED） }

  # ============================================================
  # E7 - 品质管控
  # ============================================================
...